Restructured modules and general cleanup
