--- Config 10.txt
@@ -1,25 +1,0 @@
-dir	x	y	y	incl	len	
-90.0	1.6	1.3999999999999995	0	4.0	7	
-0.0	0.8	4.6	0	4.0	7	
-0.0	0.8	3.8	0	4.0	7	
-0.0	0.8	2.9999999999999996	0	4.0	7	
-90.0	0.8	2.1999999999999993	0	4.0	7	
-90.0	0.8	1.3999999999999995	0	4.0	7	
-90.0	-1.6	1.3999999999999995	0	4.0	7	
-180.0	-0.8	4.6	0	4.0	7	
-180.0	-0.8	3.8	0	4.0	7	
-180.0	-0.8	2.9999999999999996	0	4.0	7	
-90.0	-0.8	2.1999999999999993	0	4.0	7	
-90.0	-0.8	1.3999999999999995	0	4.0	7	
-270.0	-1.6	-1.3999999999999995	0	4.0	7	
-180.0	-0.8	-4.6	0	4.0	7	
-180.0	-0.8	-3.8	0	4.0	7	
-180.0	-0.8	-2.9999999999999996	0	4.0	7	
-270.0	-0.8	-2.1999999999999993	0	4.0	7	
-270.0	-0.8	-1.3999999999999995	0	4.0	7	
-270.0	1.6	-1.3999999999999995	0	4.0	7	
-360.0	0.8	-4.6	0	4.0	7	
-360.0	0.8	-3.8	0	4.0	7	
-360.0	0.8	-2.9999999999999996	0	4.0	7	
-270.0	0.8	-2.1999999999999993	0	4.0	7	
-270.0	0.8	-1.3999999999999995	0	4.0	7
