--- pyfiles/__init__.py
@@ -1,3 +1,0 @@
-from pyfiles.PileGroupOpt import PileOptModel
-
-from pyfiles.influenceSolver import pileInfluenceRun
