+++ pyfiles/old/influenceSolver.py
