+++ pyfiles/old/pkr54Solver.py
